pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "piliq"
dynamic = ["version"]
dependencies = ["numpy", "pillow"]
authors = [{ name="cubicibo", email="55701024+cubicibo@users.noreply.github.com"},]
description = "Basic PIL-libimagequant inteface to palettize RGBA images."
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
readme = "README.md"

[metadata]
version = {attr = "package.__metadata__.__version__"}

[project.urls]
"Homepage" = "https://github.com/cubicibo/piliq"
"Bug Tracker" = "https://github.com/cubicibo/piliq/issues"